pyproject.toml
[project]
name = "lightning-whisper-mlx"
version = "0.0.10"
description = "High-performance Whisper speech-to-text and F5 text-to-speech for Apple Silicon"
requires-python = ">=3.10"
dependencies = [
    "huggingface_hub",
    "mlx",
    "numba",
    "numpy",
    "tqdm",
    "tiktoken==0.3.3",
    "scipy",
]

[project.optional-dependencies]
tts = ["f5-tts-mlx"]
diarize = ["pyannote-audio>=3.1,<5"]
dev = ["pytest"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lightning_whisper_mlx"]

[tool.pytest.ini_options]
addopts = "-m 'not slow'"
markers = [
    "slow: marks tests that download models and run E2E (run with '-m slow')",
]
